DM: Update the vcpu id type as uint16_t for vm_create_vcpu
Update the vcpu id type as uint16_t for vm_create_vcpu, this
keeps alignment with the updates for the structure acrn_create_vcpu
used by hcall_create_vcpu in the hypervisor.

In the device model, the caller is responsible for vcpu id type
conversion; vcpu id type is uint16_t for external interface in the
current implement.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ devicemodel/core/main.c
@@ -259,7 +259,7 @@ add_cpu(struct vmctx *ctx, int guest_ncpus)
 	int error;
 
 	for (i = 0; i < guest_ncpus; i++) {
-		error = vm_create_vcpu(ctx, i);
+		error = vm_create_vcpu(ctx, (uint16_t)i);
 		if (error != 0) {
 			fprintf(stderr, "ERROR: could not create VCPU %d\n", i);
 			return error;
